cosme: move to folders
